move more stuff to generic optimizer
